Convert rewards and transactions to use cents internally and doubles externally
Refactors the currency representation to maintain integers in the database (cents)
while exposing doubles (dollars) to the UI. Updates formatAsCurrency to handle
double values with proper decimal formatting.

+++ core/common/src/main/kotlin/nz/mikhailov/motiv/core/common/FormatExtensions.kt
@@ -2,4 +2,4 @@ package nz.mikhailov.motiv.core.common
 
 import kotlin.math.abs
 
-fun Int.formatAsCurrency() = "%s$%d".format(if (this < 0) "-" else "", abs(this))
+fun Double.formatAsCurrency() = "%s$%.2f".format(if (this < 0) "-" else "", abs(this))
